move import_file around
